backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-pptx==0.6.21
python-docx==0.8.11
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
